VACMS-202474 Find Forms independent deploy work
