+++ src/applications/static-pages/find-forms/actions/index.js
@@ -1,11 +1,11 @@
 import URLSearchParams from 'url-search-params';
-import recordEvent from '~/platform/monitoring/record-event';
+import recordEvent from 'platform/monitoring/record-event';
 import { fetchFormsApi } from '../api';
-import { MAX_PAGE_LIST_LENGTH } from '../containers/SearchResults';
 import {
   FETCH_FORMS,
   FETCH_FORMS_FAILURE,
   FETCH_FORMS_SUCCESS,
+  MAX_PAGE_LIST_LENGTH,
   UPDATE_HOW_TO_SORT,
   UPDATE_PAGINATION,
   UPDATE_RESULTS,
