+++ src/applications/static-pages/find-forms/components/SearchResult.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { format, parseISO, isAfter } from 'date-fns';
 import PropTypes from 'prop-types';
 import { replaceWithStagingDomain } from 'platform/utilities/environment/stagingDomains';
-import environment from '~/platform/utilities/environment';
-import recordEvent from '~/platform/monitoring/record-event';
+import environment from 'platform/utilities/environment';
+import recordEvent from 'platform/monitoring/record-event';
 import * as customPropTypes from '../prop-types';
 import { FORM_MOMENT_PRESENTATION_DATE_FORMAT } from '../constants';
 import FormTitle from './FormTitle';
