+++ src/applications/static-pages/find-forms/constants/index.js
@@ -15,3 +15,4 @@ export const FORM_MOMENT_CONSTRUCTOR_DATE_FORMAT = 'yyyy-mm-dd';
 export const FORM_MOMENT_PRESENTATION_DATE_FORMAT = 'MMMM yyyy';
 export const UPDATE_HOW_TO_SORT = 'findVAForms/UPDATE_HOW_TO_SORT';
 export const UPDATE_RESULTS = 'findVAForms/UPDATE_RESULTS';
+export const MAX_PAGE_LIST_LENGTH = 10;
